requirements.txt
matplotlib==3.8.2
numpy==1.26.4
pandas==2.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.27.1
scikit-learn==1.4.2
seaborn==0.12.2
socketclusterclient==1.3.6
stocktrends==0.1.5
streamlit==1.32.0
TA-Lib
urllib3==1.26.8
websocket-client==0.48.0 